airflow/fetchfrombb/functions/ml-cluster-transit-deserts/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
scikit-learn==1.3.*
numpy==1.*
db-dtypes==1.*